Refactor example app to look like others
