+++ example/internal/pkg/handlers/index.html
